feature/#10 use bottom nav
