+++ src/assets/styles/_index.scss
@@ -1,5 +1,6 @@
 // color
-$green: #259c4d;
+$green: #35ac5d;
 
 // style
-$bottom-nav-height: 64px;
+$max-width: 536px;
+$bottom-nav-height: 56px;
