+++ src/pages/index.vue
@@ -1,66 +1,118 @@
 <template>
-  <div class="index-wrapper">
-    <template v-if="loading">
-      <loading />
-    </template>
-    <template v-if="!loading">
+  <div class="list-wrapper">
+    <div class="list">
       <div class="header">
         <span>(ここにロゴを置くだなも)</span>
       </div>
-      <div class="main">
-        <google-login v-if="!login" />
-        <weekly-bells v-if="login" :uid="uid" @finish="childFinish" />
-        <link-button v-if="!login || !childLoading" to="/list">
-          みんなのカブ価を見てみる！
-        </link-button>
+      <simple-loading v-if="loading" />
+      <div v-if="!loading" class="main">
+        <div class="date">
+          <button @click="previous" v-show="showPrevious" class="date__previous">
+            ←
+          </button>
+          <div class="term">
+            <span class="term__text">{{ displayTerm }}</span>
+          </div>
+          <button @click="next" v-show="showNext" class="date__next">
+            →
+          </button>
+        </div>
+        <div class="person-price">
+          <div v-for="personPrice in priceList" :key="personPrice.userName">
+            <person-bell :user-name="personPrice.userName" :prices="personPrice.prices" :this-week="thisWeek" />
+          </div>
+        </div>
       </div>
-    </template>
+    </div>
+    <bottom-navigation />
   </div>
 </template>
 
 <script>
-import LinkButton from '~/components/atoms/LinkButton.vue';
-import GoogleLogin from '~/components/organisms/GoogleLogin.vue';
-import Loading from '~/components/organisms/Loading.vue';
-import WeeklyBells from '~/components/organisms/WeeklyBells.vue';
+import PersonBell from '~/components/molecules/PersonBell.vue';
+import SimpleLoading from '~/components/organisms/SimpleLoading.vue';
+import BottomNavigation from '~/components/organisms/BottomNavigation.vue';
+import { getBaseSundays, getYYYYMMDD, getTerm } from '~/domains/date/DateUtil';
 
 export default {
   components: {
-    LinkButton,
-    GoogleLogin,
-    Loading,
-    WeeklyBells,
+    PersonBell,
+    SimpleLoading,
+    BottomNavigation,
   },
   data() {
     return {
-      login: false,
-      uid: '',
+      displayDateList: [],
+      dateIndex: 0,
+      priceList: [],
       loading: true,
-      childLoading: true,
     };
   },
+  computed: {
+    baseDate() {
+      return this.displayDateList[this.dateIndex];
+    },
+    collectionName() {
+      const YYYYMMDD = getYYYYMMDD(this.baseDate);
+      return `prices${YYYYMMDD}`;
+    },
+    displayTerm() {
+      return getTerm(this.baseDate);
+    },
+    showPrevious() {
+      return this.dateIndex < 1;
+    },
+    showNext() {
+      return this.dateIndex > 0;
+    },
+    thisWeek() {
+      return this.dateIndex === 0;
+    },
+  },
   async created() {
-    await this.$auth().onAuthStateChanged((user) => {
-      if (user) {
-        this.login = true;
-        this.uid = user.uid;
-      } else {
-        this.login = false;
-      }
-      this.loading = false;
-    });
+    this.displayDateList = getBaseSundays();
+    await this.load();
   },
   methods: {
-    childFinish() {
-      this.childLoading = false;
+    async load() {
+      this.loading = true;
+      this.priceList.length = 0;
+      const snapshot = await this.$firestore.collection(this.collectionName).get();
+      snapshot.forEach((doc) => {
+        const userName = doc.data().userName;
+        const prices = [];
+        doc.data().prices.forEach((daily) => {
+          prices.push(daily.price.am);
+          prices.push(daily.price.pm);
+        });
+        this.priceList.push({
+          userName,
+          prices,
+        });
+      });
+      this.loading = false;
+    },
+    async previous() {
+      this.dateIndex++;
+      await this.load();
+    },
+    async next() {
+      this.dateIndex--;
+      await this.load();
     },
   },
 };
 </script>
 
 <style lang="scss" scoped>
-.index-wrapper {
-  padding: 16px 0 64px;
+.list-wrapper {
+  width: 100vw;
+}
+
+.list {
+  margin: 0 auto;
+  max-width: $max-width;
+  padding: 16px 6vw calc(48px + #{$bottom-nav-height});
 
   .header {
     align-items: center;
@@ -71,6 +123,43 @@ export default {
 
   .main {
     margin-top: 16px;
+
+    .date {
+      margin-top: 16px;
+      position: relative;
+      text-align: center;
+      width: 100%;
+
+      .term {
+        &__text {
+          background-color: #239dca;
+          border-radius: 8px;
+          color: white;
+          font-weight: bold;
+          padding: 4px 16px;
+        }
+      }
+
+      &__previous {
+        left: 32px;
+        position: absolute;
+        top: 2px;
+      }
+
+      &__next {
+        right: 32px;
+        position: absolute;
+        top: 2px;
+      }
+    }
+
+    .person-price {
+      margin-top: 40px;
+    }
+
+    &__link {
+      margin-top: 16px;
+    }
   }
 }
 </style>
